Fixed major problem where CartItems werent tracked by EF. Users can not add multiple products to their cart without problems.

+++ src/AuthService/Models/User.cs
@@ -6,5 +6,5 @@ public class User : IdentityUser
 {
     public string FirstName { get; set; } = string.Empty;
     public string LastName { get; set; } = string.Empty;
-    public double AccountBalance { get; set; } = 1000;
+    public decimal AccountBalance { get; set; } = 1000;
 }
